Refactor into analyzer package.
