+++ src/main/java/com/graphhopper/sna/analyzers/UnweightedGraphAnalyzer.java
@@ -22,8 +22,9 @@
  * You should have received a copy of the GNU General Public License along with
  * GraphHopper-SNA. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.graphhopper.sna.centrality;
+package com.graphhopper.sna.analyzers;
 
+import com.graphhopper.sna.centrality.BFSForCentrality;
 import com.graphhopper.sna.data.UnweightedNodeBetweennessInfo;
 import com.graphhopper.sna.data.UnweightedPathLengthData;
 import com.graphhopper.sna.progress.NullProgressMonitor;
