+++ src/main/java/com/graphhopper/sna/analyzers/WeightedGraphAnalyzer.java
@@ -22,8 +22,9 @@
  * You should have received a copy of the GNU General Public License along with
  * GraphHopper-SNA. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.graphhopper.sna.centrality;
+package com.graphhopper.sna.analyzers;
 
+import com.graphhopper.sna.centrality.DijkstraForCentrality;
 import com.graphhopper.sna.data.WeightedNodeBetweennessInfo;
 import com.graphhopper.sna.data.WeightedPathLengthData;
 import com.graphhopper.sna.progress.NullProgressMonitor;
